Replace rename() with mv()
When the directories used for containing the files and temp-files (for
uploads) are mounted in from the host machine to the container, node's
rename() function errors when trying to move files from the temp-file
directory to the destination directory once an upload is complete. (This
is probably related to the fact that rename() can't rename across
filesystems and so it somehow doesn't mesh well with volume mounting in
containers).